Documented the thermo vPLC and modified all register definitions to use the u16 type

+++ i2c/src/lib.rs
@@ -1,19 +1,19 @@
 //! Public constants used by other programs
-pub const ADDR_T1_C: usize = 0x0001;
-pub const ADDR_T2_C: usize = 0x0002;
-pub const ADDR_T3_C: usize = 0x0003;
-pub const ADDR_T4_C: usize = 0x0004;
-pub const ADDR_T5_C: usize = 0x0005;
-pub const ADDR_T6_C: usize = 0x0006;
-pub const ADDR_T7_C: usize = 0x0007;
+pub const ADDR_T1_C: u16 = 0x0001;
+pub const ADDR_T2_C: u16 = 0x0002;
+pub const ADDR_T3_C: u16 = 0x0003;
+pub const ADDR_T4_C: u16 = 0x0004;
+pub const ADDR_T5_C: u16 = 0x0005;
+pub const ADDR_T6_C: u16 = 0x0006;
+pub const ADDR_T7_C: u16 = 0x0007;
 
-pub const ADDR_TA1_C: usize = 0x0008;
-pub const ADDR_TA2_C: usize = 0x0009;
-pub const ADDR_TA3_C: usize = 0x000A;
-pub const ADDR_TA4_C: usize = 0x000B;
-pub const ADDR_TA5_C: usize = 0x000C;
-pub const ADDR_TA6_C: usize = 0x000D;
-pub const ADDR_TA7_C: usize = 0x000E;
+pub const ADDR_TA1_C: u16 = 0x0008;
+pub const ADDR_TA2_C: u16 = 0x0009;
+pub const ADDR_TA3_C: u16 = 0x000A;
+pub const ADDR_TA4_C: u16 = 0x000B;
+pub const ADDR_TA5_C: u16 = 0x000C;
+pub const ADDR_TA6_C: u16 = 0x000D;
+pub const ADDR_TA7_C: u16 = 0x000E;
 
-pub const ADDR_AMBIENT_C: usize = 0x000F;
-pub const ADDR_BARREL_KPA: usize = 0x0010;
+pub const ADDR_AMBIENT_C: u16 = 0x000F;
+pub const ADDR_BARREL_KPA: u16 = 0x0010;
